Refactored and improved the getFunctionCalls
Before I was parsing assuming that the icStub that directly follows a getFunction was the correct one.
Now it is a little be better, I assume the getFunction is used only once and I take its use as the icStub.
Everything between the two instructions is considered as part of arguments.
I also cleaned up the class to remove the attribute constants that was never used.