Fixed last bug
Seems like finding a mapping between a base version and a 2 steps optimized function is hard, even in cases where we thought it would be easy. I changed the implementation to make a copy, so that a function always exits to the previous version. This makes a lot of code obsolete and hence, I need to remove it.

Next: clean up.